scanner_bot/fx_insight_bot/pyproject.toml
[project]
name = "fx-insight-bot"
version = "0.1.0"
description = "FX Quant Insight Telegram Bot - daily market analysis inspired by BAML framework"
requires-python = ">=3.11"
dependencies = [
    "polygon-api-client>=1.13",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
    "pandas>=2.0,<3.0",
    "pyarrow>=14.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
